# Operator graph optimizer