--- a/nengo/builder/operator.py
+++ b/nengo/builder/operator.py
@@ -92,7 +92,7 @@ class Operator(object):
         return ''
 
     def _tagstr(self):
-        return ('"%s"' % self.tag) if self.tag is not None else ''
+        return (' "%s"' % self.tag) if self.tag is not None else ''
 
     @property
     def all_signals(self):