--- a/nengo/utils/graphs.py
+++ b/nengo/utils/graphs.py
@@ -98,6 +98,45 @@ def toposort(edges):
     return ordered
 
 
+def transitive_closure(edges, topo_sorted=None):
+    """Constructs the transitive closure of a directed acyclic graph (DAG).
+
+    The complexity is O(nodes + vertices).
+
+    Parameters
+    ----------
+    edges : dict
+        Dict of the form ``{a: {b, c}}`` where ``b`` and ``c`` depend on ``a``.
+        Must not contain cycles.
+    topo_sorted : sequence, optional
+        The topological sorting of the vertices. If not passed in, the
+        algorithm will do a topological sort.
+
+    Returns
+    -------
+    The transitive closure using the same data structure as `edges`: a dict
+    of the form ``{a: {b, c}}`` where ``b`` and ``c`` are nodes that either
+    directly or indirectly depend on ``a``.
+    """
+    if topo_sorted is None:
+        topo_sorted = toposort(edges)
+
+    sets = {}
+    reachables = {}
+    for vertex in reversed(topo_sorted):
+        reachables[vertex] = set(edges[vertex])
+        for edge in edges[vertex]:
+            reachables[vertex].update(reachables[edge])
+        reachables[vertex] = frozenset(reachables[vertex])
+
+        # We try to reuse existing sets as this can significantly reduce
+        # memory in some cases (which is important in the OpMergeOptimizer).
+        if reachables[vertex] in sets:
+            reachables[vertex] = sets[reachables[vertex]]
+        sets[reachables[vertex]] = reachables[vertex]
+    return reachables
+
+
 def reverse_edges(edges):
     """Reverses direction of dependence dict.