… CHANGES.rst
… docs/backend_api.rst
… examples/usage/rectified_linear.ipynb
… nengo/builder/tests/test_optimizer.py
… nengo/tests/test_learning_rules.py
… nengo/utils/tests/test_graphs.py
… nengo/utils/tests/test_stdlib.py