I updated this description with the new template. The old description can be found below.

**Description:**
This adds an optimizer to the simulator that goes through the operator graph and merges operators and their signals. This reduces the items to iterate over in slow Python (vs. fast NumPy) and gives larger blocks of sequentially aligned memory for improved CPU caching and prefetching.

**Motivation and context:**

Not everyone has a powerful GPU and can use nengo_ocl (especially at the summer school it is common to have a laptop without decent GPU). For those people this PR can significantly improve the run time.

This is also useful for running on high-performance clusters (in a serial-farming setup) like Sharcnet which mostly consist of CPU clusters and only few GPU clusters (and usually much more CPU cores than GPUs are provided).

Furthermore, when using large ensemble arrays (as in SPA, especially with spaopt #941, dot products, circular convolutions) this reduces the overhead that is introduced by many small ensembles.

**Interactions with other PRs:**
No real interactions afaik, but #941 will benefit.

**How has this been tested?**
The optimizer is enabled by default and all the tests still pass. Furthermore, I am using the optimizer on my TCM model and on @ikajic's spa-rat model. Especially, the latter one has a large arrays of small ensembles (5018 elements in a cleanup memory) and uncovered a number of memory and runtime issues that have been fixed. @xchoo tested a slightly older version of the optimizer on Spaun. Read through the comments for quotes on run time improvements and additions to the build time (quoted build time for Spaun might be lower now with recent improvements).

I did not add a lot of unit tests because for the most part it is not clear to me how to test the optimizer. It is tested through the existing tests (all those models still work) which verifies its correctness in a way, but doesn't test that it actually speed up things or isn't a no-op. Good ideas welcome.

**Where should a reviewer start?**
`optimizer.py` is the main algorithm and should be structured going from the high level to the low level of the algorithm. I tried to document it well, but it might not be sufficient to understand everything. Let me know where more comments would be helpful.

Maybe it is best to go through the code with me as it is somewhat complex.

**How long should this take to review?**

<!--- Please estimate if this PR is a quick, average, or lengthy PR. -->

<!--- Take into account both the size and complexity of the changes. -->

<!--- Also note if this is a timely PR that should be reviewed by a certain date. -->

<!--- Leave only the line that applies below: -->
- Lengthy (more than 150 lines changed or changes are complicated)

**Types of changes:**
- New feature (non-breaking change which adds functionality)

**Checklist:**

<!--- Go over all the following points. Put an `x` in all the boxes that apply. -->

<!--- If a box is not applicable, please justify below the checklist. -->

<!--- If you're unsure about any of these, don't hesitate to ask. -->

<!--- We're here to help! -->
- [x] I have read the **CONTRIBUTING.rst** document.
- [x] I have updated the documentation accordingly. (More documentation should be added, separate issue #1119.)
- [x] I have included a changelog entry.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.

**Still to do:**
#1119, but I think it should be a separate PR.
# Old description for reference

This adds an optimizer to the simulator that goes through the operator graph and merges operators and their signals. This reduces the items to iterate over in slow Python (vs. fast NumPy) and gives larger blocks of sequentially aligned memory for improved CPU caching and prefetching.

At the moment the only operator that's being merged is `SimNeurons`, but I will look and see if this can be done for other operators too.

There is a trade-off, though. Running the simulation gets faster, but the build process will be slower. The algorithmic complexity of the optimizer is O(n^2 + e) where n is the number of operators and e the number of edges in the dependency graph.

A few "benchmarks" I run:
- Simulating an ensemble array with 100 one-dimensional ensembles with 50 neurons: Pure simulation time decreases from 18 to 4 seconds (4.5x speedup! :dash: )
- `py.test -n 6` from 60s down to 51s
- `py.test -n 6 --slow` from 415s down to 274s 
- `py.test` from 83s down to 71s

I haven't tested in on any more complicated models yet, but that should definitely be done. It would be nice to have #937 implemented to easily see the effect on build time.

It probably would be a good idea to have an option to turn the optimization on and off. Probably with an rc setting, but I feel like this could be something people want to change per instantiation of the `Simulator`. Also wondering if there should be some more fine grained control which operators to merge (or if we add other optimizations which set of these should be done).

~~I think there was an issue about adding optimizations to the simulator, but I can't find it right now.~~ Found it, this addresses #778.
# TODO
- [x] Merging of views has been implemented for `Copy` in the `op-opt-view` branch with a lot of code duplication. Refactor this.
- [x] Look at view merging for other operators
- [x] Look at general merging for other operators
- [x] Add more documentation about how the optimization is performed and some of the finecky details (views get broken in the first pass, then fixed later, then view merging is done because first pass introduces more views)
- [x] Could we need more than two passes? Or in other words could view merging introduce more views because of associated non-view signals? (--> only relevant if view merging is supported for other things than `Copy`)
- [x] Add test for signal merging (this should be simple enough to unit test and is part of the public `Signal` API). Not sure if it makes sense to test the optimizer in any other way than checking that it does not break any of the existing tests. Complex interactions are hard to cover in any other way anyways. Not sure if it is worth to construct and check simple cases. Maybe that certain optimizations are done at all?
- [x] Config options
- [x] Under some circumstances the optimizer seems to require a lot of memory. I should investigate this.
- [x] Always performs a minimum of two passes even if the first pass didn't lead to a reduction in the number of operators.
- [x] This currently use the `np.stack` function in one place which requires NumPy 1.10. Replace it with an equivalent available in earlier versions. Also, add a test because that part of the code is apparently never hit in the tests (no failure on Travis-CI).