@jgosmann 2016-04-26
General note:
* The optimization algorithm is greedy in figuring out what to merge. Finding the optimal solution is equivalent to finding cliques in a graph which is NP-complete.

Important note to self to fix:
* [ ] If two operators access the same signal and both get merged with other operators, things will (probably) break.