@jgosmann 2016-04-26
Another todo:
* [ ] add some logging for debugging purposes