@jgosmann 2016-04-26
Continuing my todo list:
* [ ] Merging of views has been implemented for `Copy` in the `op-opt-view` branch with a lot of code duplication. Refactor this.
* [ ] Look at view merging for other operators
* [ ] Look at general merging for other operators
* [ ] Add more documentation about how the optimization is performed and some of the finecky details (views get broken in the first pass, then fixed later, then view merging is done because first pass introduces more views)
* [ ] Could we need more than two passes? Or in other words could view merging introduce more views because of associated non-view signals?