@jgosmann 2016-04-28
I figured out how to merge `DotInc` operations given SciPy is installed (with `bsr_matrix`) and realized how to merge even more stuff by correctly propagating memory layouts from on operation to another (though this requires somewhat more optimization passes and makes the build a bit slower). On my circular convolution test case this reduces the simulation time for 75 to 5 seconds (15x speedup! :dash: )

Still have to clean up my code quite a bit before I can commit this, though.