@hunse 2016-04-28
How fast is your convolution benchmark without combining the DotIncs?