@jgosmann 2016-04-28
18s (5227 operators instead of just 79).