@hunse 2016-04-28
That's pretty good. Looking at it, it actually just [calls gemv](https://github.com/scipy/scipy/blob/master/scipy/sparse/sparsetools/bsr.h#L750) for all the individual matrices. But the fact that it's in a tight C loop removes all the Python overhead and obviously really helps. I was kind of worried it would be using some specialized sparse matrix code that won't be as fast as BLAS for large matrices, but that's obviously not the case.

The only drawback is that all the matrices have to be the same size, but that is often true because typically we have quite a few EnsembleArrays in large models. So overall, this should really help. Nice!