@jgosmann 2016-05-04
If I remember correctly the operator count was decreased to a few hundred thousand from a couple of million for Spaun.