@xchoo 2016-05-04
Specifically, 1,334,851 operators to 245,760 operators. The build time was 2588s, and the runtime for 0.1s was 147.3s.

On the master nengo branch, the build time is typically ~1200s, and the runtime for 0.1s is ~700s