@jgosmann 2016-06-02
* [ ] Under some circumstances the optimizer seems to require a lot of memory. I should investigate this.