@jgosmann 2016-06-27
I could use a memory allocation guru ... under certain circumstances the optimizer will lead to a very high memory usage (tripled usage during the first pass). My guess right now is that this is due to memory fragmentation because a lot of small memory segments are being merged into one big memory segment which does not fit in any of the freed areas of the small memory segments.

My best idea to solve this is to limit the size of newly created memory segments. Unfortunately, I have no idea what a good limit would be.