@jgosmann 2016-06-27
Apparently, the NumPy arrays weren't the problem, but some very long strings I was constructing. Still getting a somewhat high peak memory usage under certain circumstances, but at least the memory gets freeed afterwards.