@jgosmann 2016-06-27
The high peak memory usage is caused by the transitive closure ...