@hunse 2016-06-28
I'm trying to figure out if there's overlap between this and the planners that I have in nengo_ocl. They're both trying to do a similar thing, in terms of figuring out which operators can be executed in parallel and can be combined.

It's not clear to me what part of this code is figuring out what is parallel and what is not, though. Maybe I'll need an explanation in person sometime.