@jgosmann 2016-06-28
> It's not clear to me what part of this code is figuring out what is parallel and what is not, though.

To answer that specific question: It is [this line](https://github.com/nengo/nengo/blob/f4aa0e438d4fcfb0e096d91ab4429d153023764b/nengo/builder/optimizer.py#L217) where `tc` is the transititve closure computed [here](https://github.com/nengo/nengo/blob/f4aa0e438d4fcfb0e096d91ab4429d153023764b/nengo/builder/optimizer.py#L149) (if one operator is in the transitive closure of the other, one depends on the other and they cannot be executed in parallel)

At least for now. Unfortunately, the data structure for the transitive closure can take up a lot of memory, but it seems that everything using less memory will increase the runtime.

> Maybe I'll need an explanation in person sometime.

That might be necessary for the review anyways. Though, it would be good if someone could get an understanding just by reading the code ...