@jgosmann 2016-06-30
Some more stats I collected on @ikajic's spa-rat model:
* Build time: ~9min
* Optimizer run time (with latest improvements): ~4min (without those improvements > 15min)
* Simulation time without optimizer: ~4h
* Simulation time with optimizer: ~30min
* Peak memory usage without optimizer: <6GB
* Peak memory usage with optimizer: <9GB