@jgosmann 2016-06-30
For comparison with nengo_ocl (GeForce GTX 980):
* Build time: ~7min (excluding semantic pointer generation, included in stats above)
* Planning etc ~70s
* Simulation time: ~1min
* Peak memory usage: <6GB