@jgosmann 2016-06-30
Just noticed that the optimizer makes LIF neurons run faster than direct mode (at least in some cases). :open_mouth: