@jgosmann 2016-07-11
This should be ready for review now. I updated the PR description with our template.