@jgosmann 2016-08-10
This uses `np.stack` which is not available on Sharcnet (Numpy 1.9). Should be an easy fix though. Not sure why our tests are not hitting that.