@jgosmann 2016-08-17
Moved all the code specific to the optimizer to `optimizer.py`. This seems much cleaner to me. We might want to discuss the order of function and their naming ...