@tbekolay 2016-08-25
> This seems much cleaner to me.

Agreed! I took a quick look and this is much cleaner than the last time I skimmed the diff. Some other general things I saw were that a lot of uses of classes could probably be done with functions instead, and lots of docstrings don't match [PEP-257 conventions](https://www.python.org/dev/peps/pep-0257/#multi-line-docstrings), mostly in not having a one-line summary to start. I'll look at this in more detail after I wrap up some 2.2.0 stuff -- this PR's my first priority post 2.2.0.