@jgosmann 2016-08-25
> I saw were that a lot of uses of classes could probably be done with functions instead

I think the classes add useful structure here, but I'm happy to be proven wrong.

> mostly in not having a one-line summary to start

Some complicated things are really hard to express in a single line.